requirements.txt
openai==0.27.8
python-dotenv==1.0.0
tqdm==4.65.0
pillow==11.2.1